refactor: распил base-entity — YdbRepository как ядро, persistence/relations модули
- Реализация CRUD/WHERE/шифрования/enum/lazy/hooks вынесена из YdbBaseEntity
  в src/persistence/entity-persistence.ts (YdbEntityPersistence)
- Relations/eager/many-to-many — в src/relations/entity-relations.ts
- YdbRepository — публичное ядро, объединяет persistence и relations;
  репозиторий кешируется в entity-runtime (repository-resolver)
- YdbBaseEntity — тонкий AR-фасад (~286 строк): статики делегируют
  в репозиторий; instance-helpers (toJSON, decryptField, loadRelations)
  сохранены
- NestJS forFeature: DI-провайдер репозитория возвращает тот же инстанс,
  что использует Active Record
- Публичный API не изменён; README/AGENTS.md актуализированы
