+++ AGENTS.md
@@ -23,12 +23,17 @@
 
 - `core/` — типы (`YdbPrimitive`), интерфейсы опций модуля (`YdbModuleOptions`, `YdbModuleAsyncOptions`, `YdbExecutor`), DI-токены (`constants.ts`), маппер TS→YDB, `quoteIdentifier`, `QueryOptions` (`trx`, `timeout`, `signal`, `limit`, `offset`).
 - `decorators/` — `@YdbEntity('table')`, `@YdbColumn` / `@YdbPrimaryColumn` (составной PK поддерживается), `@YdbEncrypted({ blindIndex })`, `@YdbSecurityAAD()`, `@OneToMany` / `@ManyToOne` / `@OneToOne`, `@EagerLoad([...])` (batch одним `IN (...)`, без N+1). Метаданные декораторов — copy-on-write (наследники не портят метаданные родителя).
-- `entity/base-entity.ts` — `YdbBaseEntity` (Active Record): статические `find`, `findAll` (limit по умолчанию 100, макс 1000), `count`, `save` (upsert/update по PK, составной PK поддерживается во всех CRUD-операциях), `insertMany` (батчи по 100, внутри батча группировка по набору колонок — опущенные поля не затираются NULL), `loadRelations`. PK по умолчанию — `uuid`, автогенерация через `uuid` v4. Рантайм-зависимости (executor, провайдеры шифрования) — в `entity/entity-runtime.ts` (WeakMap по классу, наследники не разделяют состояние).
+- `entity/base-entity.ts` — `YdbBaseEntity`: тонкий Active Record-фасад. Содержит только публичные статические методы (`find`, `findAll`, `count`, `save`, `insertMany`, `updateBy`, `delete`, `deleteBy`, `query`) и instance helpers (`toJSON`, `decryptField`, `decryptLazyFields`, `loadRelations`). Вся реализация делегирует в `YdbRepository`/`YdbEntityPersistence`/`YdbEntityRelations`, которые хранятся в `entity/entity-runtime.ts`.
+- `persistence/entity-persistence.ts` — `YdbEntityPersistence<T>`: ядро CRUD, WHERE-генерация, шифрование/дешифровка, blind index, enum-конвертация, timestamp-автопростановка, lifecycle hooks, `insertMany` с группировкой по колонкам, `updateBy`/`deleteBy` с RETURNING по PK.
+- `relations/entity-relations.ts` — `YdbEntityRelations<T>`: eager loading (`@EagerLoad`) одним `IN (...)`, `loadRelations`, many-to-many через join-таблицу.
+- `repository/ydb-repository.ts` — `YdbRepository<T>`: публичный DI-API, объединяет `YdbEntityPersistence` и `YdbEntityRelations`.
+- `repository/repository-resolver.ts` — `getOrCreateRepository(entityClass)`: создаёт/кеширует `YdbRepository` из deps в `entity-runtime`.
+- `entity/entity-runtime.ts` — runtime-зависимости сущности (executor, encryption/blind-index/validation провайдеры, `uuidGenerator`, кешированный `YdbRepository`). WeakMap по классу: наследники не разделяют состояние.
 - `encryption/` — интерфейсы `YdbEncryptionProvider` / `YdbBlindIndexProvider` (encrypt/decrypt с AAD, hash для blind index). `@YdbEncrypted({ lazy: true })` — ленивая дешифровка: поле дешифруется не при SELECT, а явно через `decryptField()`/`decryptLazyFields()` на инстансе; `toJSON()` требует предварительной дешифровки. **Внимание:** `Base64TestEncryptionProvider` — заглушка для тестов, реальной криптографии нет.
 - `metadata/` — `entity-metadata.ts` (сбор метаданных из Reflect, кеш на класс) и `entity-registry.ts` (глобальный реестр сущностей: `@YdbEntity` регистрирует класс при загрузке файла; используется schema sync).
 - `module/` — интеграция с NestJS:
   - `ydb-core.module.ts` — глобальный `YdbCoreModule.forRootAsync(...)` (`useFactory` / `useClass` / `useExisting`): `Driver`, `YdbExecutor` (через `query(driver)`), credentials provider по `auth_type` (невалидное значение — ошибка `Invalid YDB auth type`), опциональные encryption/blind-index провайдеры, schema sync.
-  - `ydb.module.ts` — `YdbModule.forFeature([...Entity])`: через `repository-factory.ts` сущностям инжектируется executor и провайдеры шифрования (без `forFeature` статические методы упадут с понятной ошибкой про `YdbModule.forFeature([...])` / `configureEntities([...])`). Кроме того, `forFeature` регистрирует инжектируемые `YdbRepository<Entity>` — токен получается через `getRepositoryToken(Entity)` или декоратор `@InjectRepository(Entity)`. `YdbRepository` — тонкая обёртка над статическими методами Active Record; также доступен `YdbEntityManager` (фабрика репозиториев).
+  - `ydb.module.ts` — `YdbModule.forFeature([...Entity])`: через `repository-factory.ts` сущностям инжектируется executor и провайдеры шифрования, и для каждой создаётся `YdbRepository<Entity>` (без `forFeature` статические методы упадут с понятной ошибкой про `YdbModule.forFeature([...])` / `configureEntities([...])`). `forFeature` регистрирует инжектируемые репозитории — токен получается через `getRepositoryToken(Entity)` или декоратор `@InjectRepository(Entity)`. DI-провайдер репозитория возвращает тот же инстанс `YdbRepository`, что используется Active Record. Также доступен `YdbEntityManager` (фабрика репозиториев).
 - `schema/schema-sync.ts` — генерация и проверка схемы БД (`YdbSchemaSyncer`): `sync: true` в опциях `forRootAsync` при старте создаёт недостающие таблицы (`CREATE TABLE`) и колонки (`ALTER TABLE ADD COLUMN`) по метаданным всех зарегистрированных сущностей (включая synthetic `{field}_bi` колонки blind index). Описание таблицы — через Table service `DescribeTable` (query service метаданные колонок не отдаёт). Расхождение типа колонки или PK — ошибка (в YDB не меняется); лишние колонки не удаляются (только warn). Провайдер `YDB_SCHEMA_SYNC` экспортируется — `syncer.verify(entities)` проверяет схему без изменений. Каждая сущность обязана иметь PK-колонку (`@YdbPrimaryColumn` или `uuid`), иначе sync упадёт с понятной ошибкой.
 - `transaction/transaction.manager.ts` — `YdbTransactionManager.runInTransaction(fn)`; внутри транзакции методам сущностей передаётся `{ trx }` в `QueryOptions`.
 - `credentials/auth-key-credentials-provider.ts` — обмен JWT на IAM-токен через `fetch` (без `@yandex-cloud/nodejs-sdk`), кеш с leeway 60с, защита от race condition.
