+++ src/persistence/index.ts
@@ -1,0 +1,8 @@
+export {
+  YdbEntityPersistence,
+  getEntityDbSchema,
+  hasLazyPendingCiphertext,
+  getLazyPendingFieldNames,
+  type YdbEntityConstructor,
+  type PersistenceDeps,
+} from './entity-persistence.js';
