+++ src/relations/index.ts
@@ -1,0 +1,1 @@
+export { YdbEntityRelations, type RelationsDeps } from './entity-relations.js';
