+++ src/repository/index.ts
@@ -1,3 +1,4 @@
 export { YdbRepository, type YdbEntityConstructor } from './ydb-repository.js';
 export { YdbEntityManager } from './ydb-entity-manager.js';
 export { getRepositoryToken, InjectRepository } from './repository-token.js';
+export { getOrCreateRepository } from './repository-resolver.js';
