+++ src/repository/ydb-repository.ts
@@ -1,22 +1,20 @@
 import type { QueryOptions } from '../core/query-options.js';
 import type { YdbBaseEntity } from '../entity/base-entity.js';
 import type { YdbQueryBuilder } from '../query/query-builder.js';
+import type {
+  YdbEntityPersistence,
+  YdbEntityConstructor,
+} from '../persistence/entity-persistence.js';
+import type { YdbEntityRelations } from '../relations/entity-relations.js';
 
-/**
- * Тип сущности-конструктора, совместимого с YdbBaseEntity.
- */
-export type YdbEntityConstructor<T extends YdbBaseEntity> = {
-  new (): T;
-} & typeof YdbBaseEntity;
+export type { YdbEntityConstructor };
 
 /**
  * Репозиторий для работы с конкретной сущностью YDB.
  *
- * Представляет собой DI-чистую обёртку над статическими методами
- * Active Record (YdbBaseEntity). Все вызовы делегируют в `EntityClass`,
- * поэтому рантайм-зависимости (executor, encryption/validation providers)
- * по-прежнему настраиваются один раз через `YdbModule.forFeature([...])`
- * или `configureEntities([...])`.
+ * Содержит `YdbEntityPersistence` (CRUD, шифрование, lifecycle hooks)
+ * и `YdbEntityRelations` (eager/lazy relations). Является публичным DI-API:
+ * `YdbModule.forFeature([...])` регистрирует инжектируемый `YdbRepository<Entity>`.
  *
  * Использование в NestJS:
  * ```ts
@@ -33,51 +31,74 @@ export type YdbEntityConstructor<T extends YdbBaseEntity> = {
  * ```
  */
 export class YdbRepository<T extends YdbBaseEntity> {
-  constructor(public readonly entityClass: YdbEntityConstructor<T>) {}
+  constructor(
+    public readonly entityClass: YdbEntityConstructor<T>,
+    public readonly persistence: YdbEntityPersistence<T>,
+    public readonly relations: YdbEntityRelations<T>,
+  ) {}
 
   find(where: Record<string, any>, options?: QueryOptions): Promise<T | null> {
-    return this.entityClass.find<T>(where, options);
+    return this.persistence.find(where, options);
   }
 
   findOneBy(
     where: Record<string, any>,
     options?: QueryOptions,
   ): Promise<T | null> {
-    return this.entityClass.findOneBy<T>(where, options);
+    return this.persistence.findOneBy(where, options);
   }
 
   findAll(where?: Record<string, any>, options?: QueryOptions): Promise<T[]> {
-    return this.entityClass.findAll<T>(where ?? {}, options);
+    return this.persistence.findAll(where ?? {}, options);
   }
 
   findBy(where: Record<string, any>, options?: QueryOptions): Promise<T[]> {
-    return this.entityClass.findBy<T>(where, options);
+    return this.persistence.findBy(where, options);
   }
 
   count(where?: Record<string, any>, options?: QueryOptions): Promise<number> {
-    return this.entityClass.count(where ?? {}, options);
+    return this.persistence.count(where ?? {}, options);
   }
 
   query(): YdbQueryBuilder<T> {
-    return this.entityClass.query<T>();
+    return this.persistence.query();
   }
 
   insertMany(entities: T[], options?: QueryOptions): Promise<T[]> {
-    return this.entityClass.insertMany<T>(entities, options);
+    return this.persistence.insertMany(entities, options);
   }
 
   save(entity: T, options?: QueryOptions): Promise<T> {
-    return this.entityClass.save<T>(entity, options);
+    return this.persistence.save(entity, options);
   }
 
-  delete(pkValue: string | number, options?: QueryOptions): Promise<T | null> {
-    return this.entityClass.delete<T>(pkValue, options);
+  delete(
+    pkValue: string | number | Record<string, any>,
+    options?: QueryOptions,
+  ): Promise<T | null> {
+    return this.persistence.delete(pkValue, options);
   }
 
   deleteBy(
     where: Record<string, any>,
     options?: QueryOptions,
   ): Promise<number> {
-    return this.entityClass.deleteBy(where, options);
+    return this.persistence.deleteBy(where, options);
+  }
+
+  updateBy(
+    where: Record<string, any>,
+    patch: Partial<Record<string, any>>,
+    options?: QueryOptions,
+  ): Promise<number> {
+    return this.persistence.updateBy(where, patch, options);
+  }
+
+  loadRelations(
+    items: T[],
+    relationNames: string[],
+    options?: QueryOptions,
+  ): Promise<void> {
+    return this.relations.loadRelations(items, relationNames, options);
   }
 }
